requirements.txt
fastapi
uvicorn
pdfminer.six
sentence-transformers
pinecone-client